Refactor (#7)
Why:

Current code is slightly messy and repetitive, as we were mostly understanding the main logic so far. Now the main flows are implemented, we can clean it up a bit.

How:

Add more properties to the base option type, such as isLong, isCall, isShort, digits, etc.
Change some functions' args to objects args, to make it easier to understand what is being passed in
Simplify useGetPremia. Currently this hook receives an option, gets the premia based on the option type, then returns the premia in ETH and USD. The components that use the hook then need to check what the option type is to know what value returned from the hook to use, then need to calculate the slippage. This changes the hook to only return ETH or USD premia, and return the slippage value.
In getTradeCalldata, convert the option size to a long integer.